refactor: rename state_control common/ to _common/
Rename the shared state_control library from common/ to _common/ to mark it
private, and update all importing controls (act_add, caa, cast, iti) and tests
to the new path. Pure rename plus import-path updates; no behavior change.
