+++ aisteer360/algorithms/state_control/act_add/control.py
@@ -7,16 +7,16 @@ import torch
 from transformers import PreTrainedModel, PreTrainedTokenizerBase
 
 from aisteer360.algorithms.state_control.base import StateControl
-from aisteer360.algorithms.state_control.common.estimators import SinglePairEstimator
-from aisteer360.algorithms.state_control.common.gates import AlwaysOpenGate
-from aisteer360.algorithms.state_control.common.hook_utils import (
+from aisteer360.algorithms.state_control._common.estimators import SinglePairEstimator
+from aisteer360.algorithms.state_control._common.gates import AlwaysOpenGate
+from aisteer360.algorithms.state_control._common.hook_utils import (
     get_model_layer_list,
     extract_hidden_states,
     replace_hidden_states,
 )
-from aisteer360.algorithms.state_control.common.selectors import FixedLayerSelector, FractionalDepthSelector
-from aisteer360.algorithms.state_control.common.steering_vector import SteeringVector
-from aisteer360.algorithms.state_control.common.transforms import AdditiveTransform, NormPreservingTransform
+from aisteer360.algorithms.state_control._common.selectors import FixedLayerSelector, FractionalDepthSelector
+from aisteer360.algorithms.state_control._common.steering_vector import SteeringVector
+from aisteer360.algorithms.state_control._common.transforms import AdditiveTransform, NormPreservingTransform
 
 from .args import ActAddArgs
