+++ tests/controls/test_cast.py
@@ -3,7 +3,7 @@ import torch
 
 from aisteer360.algorithms.core.steering_pipeline import SteeringPipeline
 from aisteer360.algorithms.state_control.cast.control import CAST
-from aisteer360.algorithms.state_control.common.steering_vector import SteeringVector
+from aisteer360.algorithms.state_control._common.steering_vector import SteeringVector
 from tests.utils.sweep import build_param_grid
 
 PROMPT_TEXT = (
